+++ server/controllers/base.js
@@ -29,7 +29,9 @@ class baseController {
       '/api/user/status',
       '/api/user/logout',
       '/api/user/avatar',
-      '/api/user/login_by_ldap'
+      '/api/user/login_by_ldap',
+      '/api/user/passkey/auth/options',
+      '/api/user/passkey/auth/verify'
     ];
     if (ignoreRouter.indexOf(ctx.path) > -1) {
       this.$auth = true;